Moving common utility functions out to sdk module
- Moved readCinderConf and getAPI plus supported regexes into
  dfs_sdk.scaffold module since each utility script was using these.
- Scripts can now import "getAPI" from dfs_sdk.scaffold to get access to
  all the previous functionality
